MDS-3-hadoop-spark/requirements.txt
requests==2.32.2
confluent-kafka==2.3.0
six>=1.16.0
pyspark==3.5.0
pandas==2.1.4
python-dotenv==1.0.0
SQLAlchemy==2.0.25
hdfs==2.7.3